Move label logic to label page
